+++ src/NeuralNet/Initializers/Normal.php
@@ -31,8 +31,8 @@ class Normal implements Initializer
     public function __construct(float $stddev = 0.05)
     {
         if ($stddev <= 0.0) {
-            throw new InvalidArgumentException('Standard deviation cannot be'
-                . " less than or equal to 0, $stddev given.");
+            throw new InvalidArgumentException('Standard deviation must'
+                . " be greater than 0, $stddev given.");
         }
 
         $this->stddev = $stddev;